Clean Up Code
