+++ presentation/src/main/java/com/tdtd/presentation/ui/detail/DetailAdminDialogFragment.kt
@@ -8,7 +8,7 @@ import androidx.databinding.DataBindingUtil
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.tdtd.presentation.R
 import com.tdtd.presentation.databinding.DetailAdminBottomSheetBinding
-import com.tdtd.presentation.ui.dialog.DialogConstructor
+import com.tdtd.presentation.ui.dialog.CustomDialogFragment
 import com.tdtd.presentation.util.setNavigationResult
 
 class DetailAdminDialogFragment : BottomSheetDialogFragment() {
@@ -43,7 +43,7 @@ class DetailAdminDialogFragment : BottomSheetDialogFragment() {
     private fun onClickDeleteRoom() {
         binding.deleteRoomTextView.setOnClickListener {
             setNavigationResult(getString(R.string.dialog_delete_room_button_submit))
-            val dialog = DialogConstructor(R.layout.dialog_delete_room)
+            val dialog = CustomDialogFragment(R.layout.dialog_delete_room)
             dialog.show(requireActivity().supportFragmentManager, dialog.tag)
             // TODO: 방 삭제 submit button click시 Activity 종료 + 토스트 출력 + API 호출
         }
