+++ presentation/src/main/res/layout/fragment_main.xml
@@ -121,7 +121,7 @@
                 tools:text="즐겨찾기" />
 
             <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/recyclerView"
+                android:id="@+id/mainRecyclerView"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_margin="8dp"
